SFEstunnel.spec: bump to 5.36 SFEsamba44.spec: bump to 4.4.7 SFEperl-compress-zlib.spec: remove spec file SFEperl-apache2-request.spec: add missing (Build)Requires SFEperl-parse-recdescent SFEperl-alien-rrdtool.spec: prevent build on omnios for now. Set DESTDIR or it tries installing into live system /usr/perl5/. SFEneon-gnu.spec: change (Build)Requires %{pnm_buildrequires_SUNWlibms_devel}, SUNWzlib, SUNWlexpt SUNWlexpt SFEnettle-gnu.spec: add /usr/gnu to Summary SFEmpfr.spec: disable _use_internal_dependency_generator or stumble over osdistro's gmp SFEmediainfo.spec: remove unusual bullet char from %description, don't depend on (OIH)'s wxwidgets SFElibmicrohttpd.spec: bump to 0.9.52 SFElibksba.spec: bump to 1.3.4 SFElibidn.spec: bump to 1.33 SFElibdvdnav.spec: remove patch1, we use gcc which understands -Wall SFEguile.spec: Bump to 2.0.11, make it 32/64 bit for autogen, add patch2 guile-02-1.8.8-dd_fd-d_fd.diff (OIH), change (Build)Requires %{pnm_buildrequires_SFEautomake_115} SFEgdk-pixbuf-gpp.spec: add missing %include packagenamemacros.inc, change (Build)Requires to SFEglib2-gpp SFEfilezilla.spec: bump to 3.14.1, set BuildRequires to SFEgcc-4.9 and set CC/CXX/CPP for -std=c++14 compiler, elfedit to find libstdc++.so.6.0.20 in /usr/gcc/4.9/lib SFEcheck.spec: change (Build)Requires to %{pnm_buildrequires_SUNWgsed}, %include packagenamemacros.inc SFEbitlbee.spec: bump to 3.4.2 release version SFEbinutils-gpp.spec: bump to 2.25, use gmake as "make -j" doesn't work on all osdistro SFEautoconf-gnu.spec: change (Build)Requires to %{pnm_buildrequires_SUNWgm4}, %{pnm_buildrequires_SUNWtexi}, include packagenamemacros.inc, remove dependency on SFEpostrun SFEamanda.spec: change Requires to %{pnm_requires_SUNWgnuplot}

+++ spec-files-extra/trunk/SFEbinutils-gpp.spec
@@ -14,13 +14,14 @@
 Name:                SFEbinutils-gpp
 IPS_package_name:    developer/g++/gnu-binutils
 Summary:             GNU binutils (/usr/g++)
-Version:             2.23.2
+Version:             2.25
 Source:              http://ftp.gnu.org/gnu/binutils/binutils-%{version}.tar.bz2
-Patch1:              binutils-01-bug-2495.diff
+#paused https://sourceware.org/bugzilla/show_bug.cgi?id=2495
+#Patch1:              binutils-01-bug-2495.diff
 #Patch2:              binutils-02-ld-m-elf_i386.diff
 #Patch3:              binutils-03-lib-amd64-ld-so-1.diff
 Patch4:              binutils-04-non-constant_initializer_op.diff
-Patch5:              binutils-05-lm.diff
+#Patch5:              binutils-05-lm.diff
 SUNW_BaseDir:        %{_basedir}
 BuildRoot:           %{_tmppath}/%{name}-%{version}-build
 %include default-depend.inc
@@ -32,6 +33,11 @@ BuildRoot:           %{_tmppath}/%{name}-%{version}-build
 %else
 BuildRequires: SUNWpostrun
 Requires: SUNWpostrun
+%endif
+
+%if %{omnios}
+BuildRequires: developer/gcc48
+#Requires:      runtime, installed anyways on omnios
 %endif
 
 %package devel
@@ -52,10 +58,11 @@ Requires:                %{name}
 %setup -q -c -n %name-%version
 cd binutils-%{version}
 %ifarch amd64
-%patch1 -p1 -b .patch01
+#paused https://sourceware.org/bugzilla/show_bug.cgi?id=2495
+#%patch1 -p1 -b .patch01
 %endif
 %patch4 -p1 -b .patch04
-%patch5 -p1 -b .patch05
+#%patch5 -p1 -b .patch05
 cd ..
 %ifarch i386 amd64
 cd binutils-%{version}
@@ -77,6 +84,11 @@ fi
 nlsopt=-enable-nls
 %else
 nlsopt=-disable-nls
+%endif
+
+%if %{omnios}
+#just take one of those osdistro gcc versions
+export PATH=`ls -1d /opt/gcc-4* |tail -1`/bin:$PATH
 %endif
 
 export CC=gcc
@@ -113,7 +125,7 @@ gsed -i.bak -e 's/-Werror//g' gold/configure
 
 gsed -i.bak -e 's/-Werror//g' gold/configure
 
-make -j$CPUS
+gmake -j$CPUS
 cd ..
 %endif
 
@@ -132,7 +144,7 @@ export LDFLAGS="$LDFLAGS32"
 	    --disable-static			\
 	    $nlsopt
 
-make -j$CPUS
+gmake -j$CPUS
 
 %install
 rm -rf $RPM_BUILD_ROOT
@@ -170,7 +182,7 @@ rm -rf $RPM_BUILD_ROOT%{_libdir}/%{_arch64}/%{_arch64}
 
 # Clashes with autoconf
 #rm $RPM_BUILD_ROOT%{_std_datadir}/info/standards.info
-rm $RPM_BUILD_ROOT%{_datadir}/info/standards.info
+[ -f $RPM_BUILD_ROOT%{_datadir}/info/standards.info ] && rm $RPM_BUILD_ROOT%{_datadir}/info/standards.info
 
 %clean
 rm -rf $RPM_BUILD_ROOT
@@ -230,8 +242,11 @@ rm -rf $RPM_BUILD_ROOT
 %endif
 
 %changelog
+* Tue Sep 22 2015 - Thomas Wagner
+- use gmake as "make -j$CPUS" doesn't work on all osdistro
 * Sat Feb  7 2015 - Thomas Wagner
 - exclude SUNWpostrun on IPS 
+- bump to 2.25
 * Sun Nov 10 2013 - Milan Jurik
 - bump to 2.23.2
 - fix IPS name
